Refactor 'parse_identifiers' into 'common' module
This is used across multiple SQL query types.
